Added safe mode feature. Bump version
Use :safe => true as argument option or use _safe_mode in your
custom class to ignore non existing argument.

By default, we try to parse the help output of the program by
executing program --help. If it's not satisfying, you can
customize the parsing by passing a block to the method _safe_args.